Unlink saving crash stacktrace from Sentry
